Update WebPageCodeBehind.gs
CreateForm's UI elements: UserName Input, DatePicker, Yes/No radio button, Submit selection button, and finally retrieve the list of user's available on the selected date. Instrument ui elements behavior in methods and data connection with google spreadsheet to store data. The spreadsheet has to be created prior and the spreadsheet access key has to be provided in the global constant.