Create SendEmailReminderForMoreSignups.gs
This google script can be built for the spreadsheet backend. It can be triggered x (> 24) hours before the event, so that if there is room to accommodate more, it encourages folks to signup. 2 separate sheets in a spreadsheet are used here. One is storing the folks who signed up for the event and other includes all folks to whom the email will be sent (encouraging them to signup).